2dcontext/wide-gamut-canvas/canvas-getImageData-e_srgb.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(drawSRGBThenGetImageData_e_sRGB, 'tests drawing SRGB color on e-sRGB \
canvas and getting e-sRGB image data.');

function assert_array_approx_equals(actual, expected, epsilon)
{
  assert_true(actual.length === expected.length);
  for (var i=0; i < actual.length; i++)
    assert_approx_equals(actual[i], expected[i], epsilon);
}

function drawSRGBThenGetImageData_e_sRGB() {
    var canvas = document.createElement('canvas');
    canvas.width = 10;
    canvas.height = 10;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'srgb', pixelFormat:'float16'});
    ctx.fillStyle = 'rgba(51, 102, 153, 0.8)';
    ctx.fillRect(0, 0, 10, 10);
    var pixel = ctx.getImageData(5, 5, 1, 1).dataUnion;
    // Check against the same color in e-sRGB. 0.01 protects the test against
    // color conversion deviations.
    assert_array_approx_equals(pixel, [0.2, 0.4, 0.6, 0.8], 0.01);
}
</script>
</body>
